chore: use notebooks routes
